castgesture/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pyyaml>=6.0
opencv-python>=4.8.0
pydantic>=2.0
